+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][unreleased]
 
+## [1.0.4][] - 2021-08-02
+
 - Identifier schema is not a Registry, it's a regular entity
 - Add default: `now` for datetime fields
 - Add Category kinds and default kind
@@ -26,7 +28,8 @@
 
 - Move initial implementation from metasql
 
-[unreleased]: https://github.com/metarhia/metadomain/compare/v1.0.3...HEAD
+[unreleased]: https://github.com/metarhia/metadomain/compare/v1.0.4...HEAD
+[1.0.4]: https://github.com/metarhia/metadomain/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/metarhia/metadomain/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/metarhia/metadomain/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/metarhia/metadomain/compare/v1.0.0...v1.0.1